cls/rbd: cleanup helper method visibility
